Тип данных цены изменен с float на int

+++ catalog/models.py
@@ -20,7 +20,7 @@ class Product(models.Model):
     description = models.TextField(verbose_name='Описание')
     preview = models.ImageField(upload_to='previews/', verbose_name='Превью', **NULLABLE)
     category = models.ForeignKey(Category, on_delete=models.SET_NULL, **NULLABLE)
-    price = models.FloatField(verbose_name='Цена')
+    price = models.IntegerField(verbose_name='Цена')
     created_at = models.DateTimeField(auto_now_add=True, verbose_name='Дата создания')
     updated_at = models.DateTimeField(auto_now=True, verbose_name='Дата последнего изменения')
